requirements.txt
flask>=3.0.0
mysql-connector-python
requests
requests-toolbelt
gunicorn
opencv-python-headless
